Move interaction to cookiecutter_wrapper
