+++ src/fprime/cookiecutter_templates/cookiecutter-fprime-component/hooks/pre_gen_project.py
@@ -1,4 +1,4 @@
-from fprime.util.interaction import is_valid_name
+from fprime.util.cookiecutter_wrapper import is_valid_name
 
 # Check to ensure Component Name is valid
 def verify_inputs(component_name, commands, events, telemetry, parameters):
