+++ src/fprime/util/commands.py
@@ -24,6 +24,7 @@ from fprime.fbuild.interaction import (
     new_project,
 )
 
+from fprime.util.cookiecutter_wrapper import new_component, new_port, new_deployment
 from fprime.util.code_formatter import ClangFormatter
 
 
@@ -128,9 +129,7 @@ def run_new(
         ___: unused pass through arguments
     """
     if parsed.component:
-        return new_component_fpp(build.deployment, parsed.platform, parsed.verbose, build)
-    if parsed.component_xml:
-        return new_component_xml(build.deployment, parsed.platform, parsed.verbose, build)
+        return new_component(build.deployment, parsed.platform, parsed.verbose, build)
     if parsed.port:
         return new_port(build.deployment, build)
     if parsed.deployment:
